s/double/float in Coeffs

+++ code/CAFAna/CAFAna/Prediction/PredictionInterpKernel.h
@@ -6,10 +6,10 @@ namespace ana
   {
 
     struct Coeffs{
-      Coeffs(double _a, double _b, double _c, double _d)
+      Coeffs(float _a, float _b, float _c, float _d)
         : a(_a), b(_b), c(_c), d(_d) {}
-      alignas(64) double a;
-      double b, c, d;
+      alignas(64) float a;
+      float b, c, d;
     };
 
     inline void ShiftSpectrumKernel(const Coeffs* __restrict__ fits,
